requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pyjwt
pydantic
email-validator
numpy
scikit-learn
pandas
